+++ sample/cpp/sample_clip_image.cpp
@@ -39,7 +39,7 @@ int main(int argc, char** argv) {
     printf("feature size: %d\n", feature_meta->embedding_num);
     std::vector<float> feature_vec(feature_meta->embedding_num);
     float* feature_ptr = reinterpret_cast<float*>(feature_meta->embedding);
-    for (size_t j = 0; j < feature_meta->embedding_num; j++) {
+    for (int32_t j = 0; j < feature_meta->embedding_num; j++) {
       feature_vec[j] = feature_ptr[j];
       std::cout << feature_vec[j] << " ";
     }
